requirements.txt

Flask==3.0.0
Flask-Cors==4.0.0
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
googletrans==4.0.0rc1
gunicorn==21.2.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dotenv==1.0.0
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
sniffio==1.3.0
uritemplate==4.1.1
urllib3==2.1.0
Werkzeug==3.0.1
gunicorn
pymongo==2.4.2